rename getopt.h to ithi_getopt.h, fix dependencies
